refactor(logger): consolidate to just the node logger (#4469)
This removes one of the concrete instances of the `Logger` interface that we currently maintain, the `console-logger`. This is the logger instance that was used when Stencil was run in-browser. After #4317 we no longer need this, so we can delete some code! 🎉

The change basically 1) removes that code and then 2) replaces all calls to `createLogger` with calls to `createNodeLogger`.