fix(compiler): reorder tsconfig#path transforms (#4501)
this commit reorders the transformations run for rewriting paths to
modules in import statements.

previously, modules in import statements would be rewritten based on
the contents of `tsconfig.json#paths` before any other transformation
was run. this would conflict with the `updateStencilCoreImports`
transformation if `@stencil/core` (or any of its valid subpaths) were
listed in `tsconfig.json#paths`. this would cause the wrong versions of
helpers such as `h` and `Host` to be picked up when a project used the
`--prerender` flag on a component that imported either of the above
helpers (the `lazy` output target version would be inlined to the output
rather than the `hydrate` version).

this commit removes the transformation `run-program` and places it
(conditionally) in the "before" transformers list _after_ stencil
imports are transformed.