chore(repo): minimize component starter test pack
this commit minimizes the number of files included in `npm pack` when
running the stencil component starter smoke tests. previously, files
would get added to the binary that would either be unnecessary, or would
not be representative of an actual stencil release. these include:

- `build/`, an artifact of building stencil itself. this isn't needed at
  all, and has been removed from the build archive
- `node_modules/` would get implicitly linked by `npm pack`, which get
  deleted now prior to running the command

a side effect of this work is that we now compartmentalize the building
of the stencil tarball (as opposed to generating a dev build for every
run of ci and pushing that to a registry). this makes the chances of
accidentally picking the wrong file when running `npm pack` and renaming
it lower.

STENCIL-850 Investigate Component Starter Issues in CI
STENCIL-688 Revisit Logic for Component Starter Tests