fix(runtime): override attrs set on Host with values from host element
This adds a runtime check for attributes set on the host element
instance ('from outside') before the attributes set on the `Host`
component ('from inside') are set on it. This allows developers to, for
instance, override a `role` attribute set on `Host` (see #3052) or to
accomplish something similar for any other attribute.