refactor(environment): remove IS_BROWSER_ENV and delete dead code
This removes the `IS_BROWSER_ENV` from our environment module.
Previously this was used to signal when Stencil was being used to
compile components in a browser environment.

After we removed support for that functionality in #4317 we don't need
the variable any longer, so we can assume that `IS_BROWSER_ENV` is
`false` and delete dead code accordingly.