chore(repo): minimize component starter test pack (#4520)
* chore(repo): minimize component starter test pack

this commit minimizes the number of files included in `npm pack` when
running the stencil component starter smoke tests. previously, files
would get added to the binary that would either be unnecessary, or would
not be representative of an actual stencil release. these include:

- `build/`, an artifact of building stencil itself. this isn't needed at
  all, and has been removed from the build archive
- `node_modules/` would get implicitly linked by `npm pack`, which get
  deleted now prior to running the command

a side effect of this work is that we now compartmentalize the building
of the stencil tarball (as opposed to generating a dev build for every
run of ci and pushing that to a registry). this makes the chances of
accidentally picking the wrong file when running `npm pack` and renaming
it lower.

STENCIL-850 Investigate Component Starter Issues in CI
STENCIL-688 Revisit Logic for Component Starter Tests

* refactor(compiler): add a Result union type for polymorphic returns (#4281)

This adds a new way to handle return types like those of the
`generateBuildStatus` function. See here for some previous conversation
about this:
https://github.com/ionic-team/stencil/pull/3383#discussion_r882905056

At present in `main` this function returns `d.CompilerBuildStats | {
diagnostics: d.Diagnostic[] }`:

https://github.com/ionic-team/stencil/blob/b5dc6d533f7961641e4876fe3411834884748b2f/src/compiler/build/build-stats.ts#L11-L14

We use this sort of pattern in a few places throughout the codebase,
where basically if a function could generate diagnostic messages it will
return either it's 'normal' return value _or_ an object holding those
diagnostics. We then typically have some manual checking to see which of
the two it is, or perhaps just a type assertion.

This pattern works alright, but in looking at this code before we
identified it as something we'd like to clean up in the future.

To do so, this commit introduces a new `Result` union type which is
either an `Ok` or an `Err`. It's quite obviously strongly inspired by
the `Result` enum in Rust's standard library
(https://doc.rust-lang.org/std/result/) and it lets us handle two
possible return values, one for the 'happy path' and one indicating
something didn't go right, in a principled way.

Using it looks something like this:

```ts
import { result } from "@utils";

const myFunc = (): result.Result<string, string> {
  if (failing) {
    return result.err("I failed!");
  } else {
    return result.ok("I succeeded!");
  }
}
```

Once you have a value of type `Result` there are some helper methods
similar to what you might be familiar with in Rust like a `result.map`.

* Revert "refactor(compiler): add a Result union type for polymorphic returns (#4281)"

This reverts commit e71e9b30eaba369bf02dc792dd2400abbe52af6f.

---------

Co-authored-by: Alice <alice.writes.wrongs@gmail.com>