fix(runtime): properly type color-interpolation-filter (#4530)
this commit updates the typings of
`SVGAttributes#color-interpolation-filter` to match the current spec,
which states the only accepted values are 'auto', 'sRGB' and
'linearRGB'.

this commit renames 's-rGB' to 'sRGB' and 'linear-rGB' to 'linearRGB'.
it also removes 'inherit'.

it is believed this was done by accident in a regex-based replacement of
camelCased properties to kebab-cased attributes in 667597dd