refactor(environment): remove IS_NODE_ENV and delete dead code (#4551)
* refactor(environment): remove IS_NODE_ENV and delete dead code

This removes the `IS_NODE_ENV` variable from our environment module
which previously was used to switch between browser-specific and
node-specific approaches to various things in the Stencil compilation
process.

Following the merge of #4317 we only support one runtime, node, for
compiling Stencil components, so we can assume that `IS_NODE_ENV` is
`true` and delete code accordingly.

* refactor(environment): remove IS_BROWSER_ENV and delete dead code

This removes the `IS_BROWSER_ENV` from our environment module.
Previously this was used to signal when Stencil was being used to
compile components in a browser environment.

After we removed support for that functionality in #4317 we don't need
the variable any longer, so we can assume that `IS_BROWSER_ENV` is
`false` and delete dead code accordingly.