chore(ci): pin puppeteer for node 14 starter tests
this commit pins the version of puppeteer used by the component starter
smoke tests. newer versions of puppeteer use the `??=` operator, which
node 14 doesn't know how to parse. since the component starter's
versioning is only expressed in git (and the cli doesn't know how to
download a specific version backed by a git-sha), we pin the version of
puppeteer to a known working version for node 14.