chore(ci): add v3-maintenance to test-browserstack (v3) (#4576)
this commit updates the `test-browserstack` workflow to allow for
browserstack to run when the `v3-maintenance` git branch is either a
pull request target (i.e. someone created a PR against the branch) or
when the branch is pushed to (i.e. someone merged a PR).

this commit also removes feature branch targets, as they are no longer
needed