fix(lazy): adjust the type of `defineCustomElements`
This fixes an issue with the type for `defineCustomElements` that we set
in output-lazy-loader.ts. In #4419 we changed the return type of the
function to be `void` (at runtime) but we didn't make a corresponding
adjustment to the type declaration for the function that we generate
here:

https://github.com/ionic-team/stencil/blob/7d5dc6cf5e0d2020c513cc87b6b2e5b93eece9bc/src/compiler/output-targets/output-lazy-loader.ts#L88

This change adjusts that type declaration to have a `void` return value,
instead of `Promise<void>`.

See #4589 for more context