refactor(cli): update check-version typings (#4600)
this commit updates the typings of `check-version` functions, to make it
clear that they can return `null` (in theory). in practice, i don't
believe this is possible. however, fixing that would require a breaking
change. for now, we'll fix the typings for 4.X and come back and remove
null checks for 5.x.