chore(ci): enable production ci releases (#4599)
this commit allows for production releases to be kicked off from a ci
workflow (github actions). it updates the `release-production` workflow
automate the following manual tasks:
1. running the prerelease scripts, which bundles stencil, validates the
   output, runs tests, etc.
2. commits the changes
3. pushes the release artifact/tag to npm
4. pushes the git tag to github

this commit is a stopgap to enable the rest of the team to be able to
perform releases. it is not our long term solution. rather, it fixes the
blocker that is rwaskiewicz when he goes on vacation ;-)