chore(jsdoc): add docs to component flag values (#4620)
this commit adds jsdoc to stencil's `CMP_FLAGS` enumeration, to help
clarify the usage of each variant in the enum. this is intended to help
guide the discussion around slots, scoped components, etc. as we work to
improve various bugs related to this part of the code over the next few
weeks