fix(typings): add additional transition events to DOMAttributes
this commit adds function declarations for the following events:
- [`onTransitionCancel`](https://developer.mozilla.org/en-US/docs/Web/API/Element/transitioncancel_event)
- [`onTransitionRun`](https://developer.mozilla.org/en-US/docs/Web/API/Element/transitionrun_event)
- [`onTransitionStart`](https://developer.mozilla.org/en-US/docs/Web/API/Element/transitionstart_event)

note that `onTransitionCancel` was  not natively supported by Chrome &
Edge until v87. this version falls _after_ stencil's currently supported
version of v79+ for both. this commit does not transpile code for the
event/provide a polyfill for versions of Chrome/Edge released before
2020.11.17

fixes: #4643