chore(repo): stub pull request action
this commit creates a new action that will (eventually) go on to create
pull requests for the stencil team when it is time to release a new
version of the library. the (eventual) intent is for stencil engineers
to run this workflow, which will generate a pull request that includes:
- version bumps in package.json & package-lock.json
- changelog, notice, and license generation

this commit currently only creates a new branch and commit. the branch
includes the version string, as well as the run/attempt in github
actions to tie us back to the run that created it.

at this point, the creation of the pull request is not included. the
intended library to use
(https://github.com/peter-evans/create-pull-request) does not work on
push/pull_request events in GHA without additional permissions. by
adding this action, we can test it without generating a new token.