chore(repo): split internal and user-facing jest config
this commit separates stencil's internal testing jest configuration
(used when running unit tests on the stencil project itself), and the
jest configuration used by stencil projects (when a stencil user runs
`npm t` in their project).

with this commit, we remove the inheritance on the jest configuration
that we bundle with stencil for end users. in order to ensure that
stencil's internal tests continue to function, the internal jest config
has been updated to have near parity with the external one. the intent
of this commit is _not_ to optimize the jest config, but rather separate
the two so that we can begin to bring support for newer versions of jest
to end users without dragging out the process (i.e. upgrading ourselves)