chore(declarations): remove unused types
this commit removes several types that are unused in the stencil
codebase. these have previously gone undetected due to their being
exported by `src/declarations/index.ts`.

these types are not used within the stencil codebse. _however_ it is
possible that a consumer is reaching into our internal type declarations
to use these types. those that are in this file are assumed to be
'private' (and therefore volatile). should a bug report come in
regarding one of these types, the recommeneded recourse is that the team
investigate/determine on a per-issue basis whether the type should be
restored in `stencil-public.ts` or re-added to this file and deprecated
(and removed in v5)