feat(config): add experimentalSlotFixes config value (#4652)
This adds a new configuration option inside of the `extras` object
called `experimentalSlotFixes`. If this is set to `true` then all slot
fix configuration options will be automatically set to `true` as well.

Furthermore, when new slot fixes are added in the future they'll be
automatically enabled if you have this flag enabled.