refactor(compiler): create type for static getters (#4738)
This is a small refactor which turns a `Set<string>` into a
`ReadonlyArray<string>` and pushes a bit more information into the type
system about how 'static getters' are created during the compilation
process.

If you're unfamiliar with this, when Stencil is compiling a component
it finds any properties which are decorated with our custom decorators
(e.g. `@Prop` and so on) and transforms the TypeScript
`ClassDeclaration` node to 1) remove the syntax nodes for the decorated
property and 2) add a `static get () { ... }` member to the class
declaration.

These static getters are then used in various ways later in the
compilation process. Generally the values stuck in them are parsed into
the `ComponentCompilerMeta`, then the `dist-custom-elements` output
target removes them entirely, and lazy-based builds (like `www`) use
them at runtime.

Right now, unfortunately, these are somewhat stringly typed! Before this
change we did not have a type-level assurance that the names being used
for these static getters were consistent. This change moves the needle
on this a bit, although it's possible that the new type (`StencilStaticGetter`)
could be used more widely.

STENCIL-856