chore(snc): bring type of string props on `Diagnostic` inline with use (#4737)
* chore(snc): bring type of string props on `Diagnostic` inline with use

This interface has two properties on it, `absFilePath` and
`relFilePath`, which are habitually instantiated as `null` throughout
the codebase but were previously typed `string | undefined`. This
changes the type on the interface declaration to be `string | undefined
| null` to match the interface to the usage.

This technically constitutes a breaking change, since it is modifying
the type of a public interface, but in practice it should not be a
problem since the type is just being widened (and widened to what it
actually will be at runtime in a lot of cases).

* keep the type as it was, change usage instead

* fix tests