refactor(compiler): create a util function for name of a property decl
Pull some code out of `convert-decorators.ts` into a helper function in
our transformer utility module for getting the name of a property
declaration as a string. If you had a class like this:

```ts
class {
  foo: "bar";
}
```

this utility would let you get the string `"foo"` given the TypeScript
syntax tree node corresponding to the `foo: "bar";` line in the source.