chore(snc): resolve snc in test init (#4770)
this commit resolves a handful of strict null checks violations in
`testing.ts`. most violations involved widening the type of some
variable to explicitly state it could be null. this ought to be a safe
operation, as we were already assigning `null` to these variables
already - the change only aligns the type system with what was
originally happening at runtime.

an extraneous call to `getAppStyleUrl` was removed as well. there are no
known side effects that occur as a result of that function call, which
ought to make this refactor safe