test(hydrate): add a regression test for hydrating SVG children (#4783)
This adds a regression test for the issue reported in #4278, which
doesn't currently occur on `main` but does occur in the branch
associated with #2938 (a PR containing a bunch of changes for fixing
bugs with slots).

This regression test will give us a bit more safety as we go to add more
slot behavior fixes in the future.