chore(release): ensure correct vermoji baked into distributable
this commit fixes a bug in our release scripts where the vermoji
generated in the prerelease step of the process would not be properly
propagated to the release step. when the entire process was manual, we'd
do it all in one fell swooop, writing a tiny file to disk and reading
that.

in lieu of that, we can pull the vermoji off of the head of the
CHANGELOG.md! this ensures consistency between the two - running
`stencil info` and looking at the changelog should show the same emoji.

note that this change intentionally does not get applied to the
`getOptions` function for releases. since we still technically support
manual releases, adding the change there for vermoji generation would
negatively impact vermoji generation