chore(repo): add karma test config to renovate
this commit updates the renovate configuration to prevent renovate from
bumping node, node typings, and npm to versions that we do not want to
run against at this time. historically, we've tested against the oldest
LTS (even if it's not supported by Node anymore, just so long as we
support it). this helps continue that.