chore(ts): enable strictBindCallApply ts compiler flag
This turns on the `strictBindCallApply` typescript compiler option and
then makes a minimal set of changes to get the build to pass with that
option on.

From [the TypeScript docs](https://www.typescriptlang.org/tsconfig#strictBindCallApply):

> When set, TypeScript will check that the built-in methods of functions
  `call`, `bind`, and `apply` are invoked with correct argument for the
  underlying function[...] Otherwise, these functions accept any
  arguments and will return `any`

Without this option doing a `.call` on a function was basically a big
type hole! Not good!