chore(changelog): reconcile changelog emoji (#4837)
this commit fixes the changelog emojis for multiple stencil releases.
when we moved to automated releases, there was a bug that would not
propagate the emoji generated in the prerelease step to the release
step. this caused a new emoji to be generated and baked into the
distributable that did not match the one in the changelog. since stencil
releases are immutable, we choose to go back and edit the changelog to
match the emojis that were baked in as a part of the release process.