chore(test): remove old jest todos
remove a handful of TODOs from the Jest 27 and under section of the
codebase. these TODOs were originally written when we planned to have to
continue to maintain the internally used version of Jest and the one
users use, keeping them in sync. now that these two versions of jest are
independent, we can remove these TODOs (as the Jest 27 and under code is
not expected to be changed for Jest 28).