chore(renovate): fix volta karma node management (#4862)
this commit attempts to fix a renovate configuration issue where we wish
to ignore npm, node and `@types/node` from getting bumped in the karma
subdirectory of the project. currently, renovate is trying to bump
these. update `matchPackageNames` to `matchDepNames` and the path to
`package.json` for node/npm to see if that picks up correctly. the node
typings will continue to use `matchPackageNames`, and have therefore
been moved to their own package rule