refactor(jest): make jest presets version-specific
this commit updates the existing infrastructure for jest presets to move
them under version-specific directories (although there's just one such
directory at this time).

in doing so, the `jest-preset.js` file that was previously housed under
the `scripts/` directory now has the same dispatch behavior as the other
javascript files in the same directory (`git mv` was used to preserve
history of the original contents). this allows the correct version of
`jest-preset` to be selected at runtime when a user invokes jest.

`jest-preset.ts` contains the original preset with the following
changes:
1. it no longer uses CommonJS exports to make loading the file at
   runtime easier and compile-time compliant. it is worth noting that
   the contents of this file, in their previous location, would simply
   be copied to the output directory. now, they are run through the
   typescript compiler and placed in `testing/index.js` after the
   compiler has been built
2. `path` is now imported with an `import` statement, rather than a
   `require` statement as a consequence of point 1.

the resulting directory structure looks something like:

```diff
  ├── scripts/bundles/helpers/jest/
  │   └── jest-preset.js                      <- Now has same dispatch behavior as other files in this directory
  ├── src/testing/jest/
  │   └── jest-27-and-under
  │       ├── jest-config.ts
  │       ├── jest-environment.ts
  │       ├── jest-facade.ts
  │       ├── jest-preprocessor.ts
+ │       ├── jest-preset.ts                  <- Contains preset previously in `scripts/`, now as a module
  │       ├── jest-runner.ts
  │       ├── jest-screenshot.ts
  │       ├── jest-serializer.ts
  │       ├── jest-setup-test-framework.ts
  │       ├── node_modules/
  │       ├── package.json
  │       └── test
  ├── jest-apis.ts
  ├── jest-facade.ts
  ├── jest-stencil-connector.ts
  └── package.json
```

the jest facade interface and it's related impl have been updated to return
the correct jest preset