chore(repo): tmp pin unit test node version (#4900)
Unit tests in CI started to fail sporadically earlier this week for both
Windows and Ubuntu on Node 20. Hoping this was sporadic, we did not
investigate immediately. Starting this morning, no CI workflows
containing our unit tests on Node 20 were passing, blocking us from landing
code.

Looking across all CI workflows that ran in the past week/week and a
half, splitting them by those that passed and those that failed
It appears the issues that plagued us during the Node 20.5 upgrade
where Jest fails when finding a file on disk (or writing it) are back

In our unit tests, we always pull the latest minor of the LTS versions of
Node we support (16. 18, 20). We’ve started getting Node 20.8 in CI,
hence the failure.

In so far as I can tell, this is only Jest test related, and does not
affect the functionality of Stencil. Pin our unit tests to 20.7 for now.

Relates to: https://github.com/ionic-team/stencil/commit/dd1fe1d0b01723ef4d88f5eb24d34eae0cf7ac1c