chore(snc): widen vnode variable type in update-element.spec
this commit widens the types of multiple VNode variable instances to
`VNode | null` from `VNode`. in each of thsee usages, an old VNode
instance is instantiated as `null` for testing purposes to call
`updateElement`. since the old VNode can be `null` according to the
`udpateElement` function signature, this is considered to by a safe
operation (i.e. does not adversely affect the tests).