chore(build): remove unnecessary edits of typescript source
This removes three edits to the TypeScript source that we make during
the Rollup build in our typescript source plugin. The edits were to:

- remove `getDefaultLibFilePath`
- remove `enableCPUProfiler`
- remove `disableCPUProfiler`

These edits were previously made in order to facilitate support for
running Stencil in the browser (in-browser compilation support) and
although they were tagged with the JIRA ticket for removing that feature
from Stencil these particular lines were overlooked.